Move to tests to MultiparentCasperDeploySpec
